[ADD/#76] 구매 정보 확인 API domain 모듈 설정
