+++ domain/src/main/kotlin/co/orange/domain/entity/response/BuyProgressModel.kt
@@ -1,16 +1,11 @@
 package co.orange.domain.entity.response
 
-import co.orange.domain.enums.OrderStatus
-
-data class BuyDetailModel(
-    val orderId: String,
-    val orderStatus: OrderStatus,
+data class BuyProgressModel(
+    val itemId: String,
     val productName: String,
     val imgUrl: String,
     val originPrice: Int,
-    val sellerNickname: String,
     val addressInfo: List<AddressInfoModel>,
-    val paymentInfo: List<PaymentInfoModel>,
     val discountPrice: Int,
     val charge: Int,
     val totalPrice: Int,
