+++ domain/src/main/kotlin/co/orange/domain/entity/response/SellProgressModel.kt
@@ -1,0 +1,7 @@
+package co.orange.domain.entity.response
+
+data class SellProgressModel(
+    val itemId: String,
+    val productName: String,
+    val originPrice: Int,
+)
