+++ domain/src/main/kotlin/co/orange/domain/repository/BuyRepository.kt
@@ -1,0 +1,7 @@
+package co.orange.domain.repository
+
+import co.orange.domain.entity.response.BuyProgressModel
+
+interface BuyRepository {
+    suspend fun getButProgressData(productId: String): Result<BuyProgressModel>
+}
