+++ presentation/src/main/java/co/orange/presentation/buy/progress/BuyProgressViewModel.kt
@@ -2,7 +2,7 @@ package co.orange.presentation.buy.progress
 
 import androidx.lifecycle.ViewModel
 import co.orange.domain.entity.response.AddressInfoModel
-import co.orange.domain.entity.response.BuyInfoModel
+import co.orange.domain.entity.response.BuyProgressModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
@@ -15,7 +15,7 @@ class BuyProgressViewModel
         var productId: String = ""
 
         val mockBuyInfo =
-            BuyInfoModel(
+            BuyProgressModel(
                 "상품이름은 한줄로만 보여줄거에야야야야야야",
                 "https://github.com/Marchbreeze/Marchbreeze/assets/97405341/cd2c0454-92b4-41e7-ae2f-319f83e2426f",
                 24000,
