+++ presentation/src/main/java/co/orange/presentation/sell/finished/SellFinishedViewModel.kt
@@ -1,7 +1,7 @@
 package co.orange.presentation.sell.finished
 
 import androidx.lifecycle.ViewModel
-import co.orange.domain.entity.response.SellInfoModel
+import co.orange.domain.entity.response.SellProgressModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
@@ -14,7 +14,7 @@ class SellFinishedViewModel
         var itemId: Long = -1
 
         val mockSellInfo =
-            SellInfoModel(
+            SellProgressModel(
                 "123e4567-e89b-12d3-a456-426614174000",
                 "딴지 키링 세트",
                 9000,
