+++ presentation/src/main/java/co/orange/presentation/sell/info/SellInfoViewModel.kt
@@ -3,7 +3,7 @@ package co.orange.presentation.sell.info
 import androidx.lifecycle.ViewModel
 import co.orange.domain.entity.response.AddressInfoModel
 import co.orange.domain.entity.response.PaymentInfoModel
-import co.orange.domain.entity.response.SellDetailModel
+import co.orange.domain.entity.response.SellInfoModel
 import co.orange.domain.enums.ItemStatus
 import co.orange.domain.enums.OrderStatus
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -18,7 +18,7 @@ class SellInfoViewModel
         var productId: Long = -1
 
         val mockSellInfo =
-            SellDetailModel(
+            SellInfoModel(
                 "1",
                 ItemStatus.ON_SALE,
                 "상품이름은 한줄로만 보여줄거에야야야야야야",
